{"schema":"decodex/commit/1","summary":"split autonomy proposal radar analysis and hero style modules","authority":"manual"}
