+++ apps/decodex/src/state.rs
@@ -78,9 +78,8 @@ use runtime_records::{
 	RunControlChannelRecord, TimestampParts, WorktreeMappingRecord,
 };
 use runtime_row_parsers::{
-	compare_autonomy_proposal_runtime_records, compare_autonomy_signal_runtime_records,
-	compare_decision_contract_runtime_records, compare_execution_program_runtime_records,
-	compare_linear_execution_event_runtime_records,
+	compare_autonomy_signal_runtime_records, compare_decision_contract_runtime_records,
+	compare_execution_program_runtime_records, compare_linear_execution_event_runtime_records,
 	compare_private_execution_event_runtime_records, compare_program_intake_plan_records,
 	compare_program_issue_mapping_records, compare_recent_autonomy_proposal_runtime_records,
 	compare_recent_autonomy_signal_runtime_records, parse_linear_execution_event_unix,
