+++ apps/decodex/src/state/store.rs
@@ -34,9 +34,8 @@ use crate::{
 		ProgramIssueMappingRecord, ProjectRegistration, ProtocolActivitySummary,
 		ReviewLifecycleRecord, ReviewPolicyCheckpoint, StateData, acquire_shared_lock_coordinator,
 		apply_derived_program_intake_state, clear_close_on_exec,
-		compare_autonomy_proposal_runtime_records, compare_autonomy_signal_runtime_records,
-		compare_decision_contract_runtime_records, compare_execution_program_runtime_records,
-		compare_linear_execution_event_runtime_records,
+		compare_autonomy_signal_runtime_records, compare_decision_contract_runtime_records,
+		compare_execution_program_runtime_records, compare_linear_execution_event_runtime_records,
 		compare_private_execution_event_runtime_records, compare_program_intake_plan_records,
 		compare_program_issue_mapping_records, compare_recent_autonomy_proposal_runtime_records,
 		compare_recent_autonomy_signal_runtime_records, dispatch_slot_lock_path,
