+++ site/src/styles/hero.css
@@ -1,338 +1,5 @@
-.official-site {
-  min-height: 100vh;
-  background:
-    linear-gradient(180deg, transparent 0, transparent 42rem, var(--bg) 42rem),
-    var(--bg);
-}
-
-.official-hero {
-  position: relative;
-  --hero-shift-x: 0px;
-  --hero-shift-y: 0px;
-  display: grid;
-  grid-template-rows: auto minmax(0, 1fr) auto;
-  min-height: min(43rem, 100vh);
-  overflow: hidden;
-  background: var(--hero-background);
-  color: var(--stage-ink);
-}
-
-.official-hero::before {
-  content: "";
-  position: absolute;
-  inset: 0;
-  z-index: 0;
-  background: var(--hero-veil);
-  pointer-events: none;
-}
-
-.official-hero::after {
-  content: "";
-  position: absolute;
-  inset: 0;
-  z-index: 1;
-  background:
-    linear-gradient(var(--hero-grid-line) 1px, transparent 1px),
-    linear-gradient(90deg, var(--hero-grid-line) 1px, transparent 1px);
-  background-size: 2.4rem 2.4rem;
-  mask-image: linear-gradient(180deg, rgba(0, 0, 0, 0.68), transparent 78%);
-  pointer-events: none;
-  transform: translate3d(var(--hero-shift-x), var(--hero-shift-y), 0);
-  transition: transform 220ms ease-out;
-}
-
-.site-header {
-  position: relative;
-  z-index: 2;
-  display: grid;
-  grid-template-columns: auto minmax(0, 1fr) auto;
-  gap: 1rem;
-  align-items: center;
-  width: min(100%, 82rem);
-  margin: 0 auto;
-  padding: 1.1rem 1.15rem;
-}
-
-.site-logo {
-  display: inline-flex;
-  width: fit-content;
-  align-items: center;
-}
-
-.site-logo__image {
-  display: block;
-  width: auto;
-  height: 2.35rem;
-  transition:
-    filter 180ms ease,
-    transform 180ms ease;
-}
-
-.site-logo:hover .site-logo__image,
-.site-logo:focus-visible .site-logo__image {
-  filter: drop-shadow(0 0 14px rgba(182, 241, 58, 0.2));
-  transform: translateY(-0.05rem);
-}
-
-.site-nav {
-  display: none;
-  align-items: center;
-  justify-content: center;
-  gap: 1rem;
-  min-width: 0;
-}
-
-.site-nav a {
-  position: relative;
-  color: var(--stage-soft);
-  font-size: 0.84rem;
-  font-weight: 740;
-  line-height: 1;
-  padding-block: 0.46rem;
-}
-
-.site-nav a::after {
-  content: "";
-  position: absolute;
-  right: 0;
-  bottom: 0.2rem;
-  left: 0;
-  height: 1px;
-  background: var(--lime);
-  opacity: 0;
-  transform: scaleX(0);
-  transform-origin: right;
-  transition:
-    opacity 180ms ease,
-    transform 180ms ease;
-}
-
-.site-nav a:hover,
-.site-nav a:focus-visible {
-  color: var(--stage-ink);
-}
-
-.site-nav a:hover::after,
-.site-nav a:focus-visible::after {
-  opacity: 1;
-  transform: scaleX(1);
-  transform-origin: left;
-}
-
-.theme-toggle {
-  display: inline-flex;
-  align-items: center;
-  gap: 0.18rem;
-  width: fit-content;
-  padding: 0.18rem;
-  border: 1px solid var(--stage-control-border);
-  border-radius: 0.42rem;
-  background: var(--stage-control-bg);
-  backdrop-filter: blur(16px);
-  transition:
-    border-color 180ms ease,
-    background-color 180ms ease,
-    box-shadow 180ms ease;
-}
-
-.theme-toggle__button {
-  min-height: 1.8rem;
-  padding: 0.35rem 0.7rem;
-  border: 0;
-  border-radius: 0.28rem;
-  background: transparent;
-  color: var(--stage-muted);
-  cursor: pointer;
-  font-size: 0.72rem;
-  font-weight: 760;
-  line-height: 1;
-}
-
-.theme-toggle__button:hover {
-  background: var(--stage-control-hover);
-  color: var(--stage-ink);
-}
-
-.theme-toggle__button[aria-pressed="true"] {
-  background: var(--stage-selected-bg);
-  box-shadow: 0 0 0 1px rgba(255, 255, 255, 0.08);
-  color: var(--stage-selected-ink);
-}
-
-.hero-grid {
-  position: relative;
-  z-index: 2;
-  display: grid;
-  grid-template-columns: minmax(0, 1fr);
-  gap: 1.4rem;
-  align-items: center;
-  width: min(100%, 82rem);
-  margin: 0 auto;
-  padding: 2.2rem 1.15rem 1.8rem;
-}
-
-.hero-copy {
-  display: grid;
-  gap: 1.05rem;
-  align-content: center;
-  min-width: 0;
-}
-
-.hero-copy h1 {
-  color: var(--stage-ink);
-  font-size: clamp(4.25rem, 13vw, 9.5rem);
-  font-weight: 850;
-  line-height: 0.82;
-}
-
-.hero-copy p {
-  max-width: 47rem;
-  color: var(--stage-soft);
-  font-size: clamp(1rem, 1.55vw, 1.24rem);
-  line-height: 1.62;
-}
-
-.hero-actions {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 0.55rem;
-}
-
-.hero-actions a {
-  position: relative;
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  min-height: 2.18rem;
-  padding: 0.52rem 0.72rem;
-  border: 1px solid var(--stage-control-border);
-  border-radius: 0.28rem;
-  color: var(--stage-ink);
-  font-size: 0.84rem;
-  font-weight: 780;
-  line-height: 1;
-  will-change: transform;
-}
-
-.hero-actions a:first-child {
-  border-color: color-mix(in srgb, var(--lime) 84%, transparent);
-  background: color-mix(in srgb, var(--lime) 12%, transparent);
-  color: var(--stage-ink);
-}
-
-.hero-actions a:hover,
-.hero-actions a:focus-visible {
-  border-color: var(--stage-action-hover-border);
-  background: var(--stage-action-hover-bg);
-  color: var(--stage-action-hover-ink);
-  transform: translateY(-0.08rem);
-}
-
-.hero-actions a:first-child:hover,
-.hero-actions a:first-child:focus-visible {
-  border-color: var(--stage-primary-action-hover-bg);
-  background: var(--stage-primary-action-hover-bg);
-  color: var(--stage-primary-action-hover-ink);
-}
-
-.hero-actions a[data-pressed="true"],
-.appcast-download__primary[data-pressed="true"],
-.appcast-download__version[data-pressed="true"],
-.doc-link[data-pressed="true"] {
-  transform: translateY(0) scale(0.985);
-}
-
-.hero-panel {
-  position: relative;
-  z-index: 20;
-  display: grid;
-  gap: 1rem;
-  align-content: end;
-  width: min(100%, 24rem);
-  justify-self: stretch;
-  padding: 1rem;
-  border: 1px solid var(--stage-control-border);
-  border-radius: 0.5rem;
-  background: var(--stage-glass-bg);
-  backdrop-filter: blur(18px);
-}
-
-.hero-panel__logo {
-  display: flex;
-  justify-content: flex-end;
-}
-
-.hero-panel__logo img {
-  width: 4.5rem;
-  height: 4.5rem;
-  border-radius: 1rem;
-  box-shadow: 0 18px 42px rgba(20, 32, 52, 0.2);
-}
-
-.hero-panel__copy {
-  display: grid;
-  gap: 0.4rem;
-  min-width: 0;
-  text-align: right;
-}
-
-.hero-panel__copy span,
-.hero-bottom span,
-.system-panel__lead span,
-.system-loop span,
-.command-panel__copy span,
-.doc-link strong,
-.appcast-download__kicker {
-  color: var(--muted);
-  font-size: 0.72rem;
-  font-weight: 780;
-  line-height: 1.2;
-  text-transform: uppercase;
-}
-
-.official-hero .hero-panel__copy span,
-.official-hero .hero-bottom span,
-.official-hero .appcast-download__kicker {
-  color: var(--stage-muted);
-}
-
-.hero-panel__copy strong {
-  color: var(--stage-ink);
-  font-size: 1.22rem;
-  font-weight: 820;
-  line-height: 1.12;
-}
-
-.hero-bottom {
-  position: relative;
-  z-index: 2;
-  display: grid;
-  grid-template-columns: repeat(2, minmax(0, 1fr));
-  width: min(100%, 82rem);
-  margin: 0 auto;
-  padding: 0 1.15rem 1.2rem;
-}
-
-.hero-bottom div {
-  display: grid;
-  gap: 0.25rem;
-  min-width: 0;
-  padding: 0.78rem 0.65rem;
-  background: var(--stage-glass-bg);
-  backdrop-filter: blur(16px);
-  transition:
-    background-color 180ms ease,
-    transform 180ms ease;
-}
-
-.hero-bottom div:hover {
-  background: var(--stage-glass-bg-hover);
-  transform: translateY(-0.08rem);
-}
-
-.hero-bottom strong {
-  color: var(--stage-ink);
-  font-size: 1.08rem;
-  font-weight: 820;
-  line-height: 1;
-}
+@import "./hero/shell.css";
+@import "./hero/header.css";
+@import "./hero/copy.css";
+@import "./hero/panel.css";
+@import "./hero/bottom.css";
